Move isv activation steps to apex-debugger extension.
