+++ packages/salesforcedx-vscode-apex-debugger/package.json
@@ -23,6 +23,7 @@
   "categories": ["Debuggers"],
   "dependencies": {
     "@salesforce/salesforcedx-apex-debugger": "44.18.0",
+    "@salesforce/salesforcedx-utils-vscode": "44.18.0",
     "vscode-debugprotocol": "1.28.0",
     "vscode-extension-telemetry": "0.0.17"
   },
